Fix lua block handling and lua dimming
There was a logic error when handling the lua block search if the
start/end tags were previously inside the same region
syntax-propertize is handling right now.

Might not be necessary, but I removed dimming overlay optimization,
and instead opt for removing all dim overlays when ranges are updated.
Which might be slower, but as pico8 files has very little code, it
should not be a problem.