Add support for alternate palette
Previously when rendering labels that had been rendered with alternate
palette colours Emacs would crash (on macOS). With the alternate colours
added it no longer crashes and renders the labels correctly. The colour
values are based on the reference palette on the sidebar of the pico8
subreddit (which differs from the default palette colours here but it
doesn't seem worth changing).