feature: Add custom for font and column fill.
This commit adds two customs: pico8-use-font and pico8-set-column-fill.
The first is disabled by default; the second is enabled by default.

I find using the pico-8 font pretty pleasant because I'm often switching
back and forth between Emacs and pico-8. Seeing the code go from
upper-case to lower-case is jarring from my perspective.

If pico8-use-font is enabled and the font isn't available. It'll report
a message with a link to where the font may be downloaded.

Thank you for this great package I've been using for years.